+++ src/router/routes/tyreModels.js
@@ -1,0 +1,7 @@
+export default [
+  {
+    path: '/tyre-models',
+    name: "tyre-models",
+    component: () => import('@/views/tyre_models/TyreModelList.vue'),
+  },
+]